+++ src/routes/apply/+page.svelte
@@ -11,7 +11,6 @@
 
 	let debounceTimer: ReturnType<typeof setTimeout> | undefined;
 	let saveButton: HTMLButtonElement;
-	let saveButtonText = typeof form === 'string' ? form : 'Save';
 
 	let confirmAction = '';
 	let dialog: HTMLDialogElement;
@@ -131,7 +130,7 @@
 			}}
 			on:input={() => {
 				saveButton.disabled = true;
-				saveButtonText = 'Autosaving...';
+				saveButton.textContent = 'Autosaving...';
 				if (debounceTimer !== undefined) {
 					clearTimeout(debounceTimer);
 				}
@@ -139,7 +138,7 @@
 					debounceTimer = undefined;
 					applicationForm.requestSubmit();
 					saveButton.disabled = false;
-					saveButtonText = 'Save';
+					saveButton.textContent = 'Save and finish later';
 				}, 1000);
 			}}
 			autocomplete="off"
@@ -147,7 +146,7 @@
 			{#each data.questions as question}
 				<div class={'question ' + question.type.toLowerCase()}>
 					<label for={question.id}>
-						{question.label}
+						<SvelteMarkdown source={question.label} isInline />
 						{#if question.required}*{/if}
 					</label>
 					<!-- svelte-ignore a11y-autofocus -->
@@ -160,6 +159,7 @@
 							placeholder={question.placeholder}
 							autofocus={question.id === focusedQuestionId}
 							on:focus={() => (focusedQuestionId = question.id)}
+							required={question.required}
 						/>
 					{:else if question.type === 'PARAGRAPH'}
 						<textarea
@@ -169,6 +169,7 @@
 							placeholder={question.placeholder}
 							autofocus={question.id === focusedQuestionId}
 							on:focus={() => (focusedQuestionId = question.id)}
+							required={question.required}
 						/>
 					{:else if question.type === 'NUMBER'}
 						<input
@@ -182,6 +183,7 @@
 							placeholder={question.placeholder}
 							autofocus={question.id === focusedQuestionId}
 							on:focus={() => (focusedQuestionId = question.id)}
+							required={question.required}
 						/>
 					{:else if question.type === 'DROPDOWN'}
 						<select
@@ -190,6 +192,7 @@
 							bind:value={application[question.id]}
 							autofocus={question.id === focusedQuestionId}
 							on:focus={() => (focusedQuestionId = question.id)}
+							required={question.required}
 						>
 							<option value="">Select...</option>
 							{#each question.options as option}
@@ -204,6 +207,7 @@
 							checked={Boolean(application[question.id])}
 							autofocus={question.id === focusedQuestionId}
 							on:focus={() => (focusedQuestionId = question.id)}
+							required={question.required}
 						/>
 					{:else if question.type === 'MULTISELECT'}
 						<Select
@@ -225,8 +229,20 @@
 				</div>
 			{/each}
 
-			<button bind:this={saveButton}>{saveButtonText}</button>
-			<button id="submit" formaction="?/finish">Submit Application</button>
+			<div id="actions-container">
+				<div id="actions">
+					<button bind:this={saveButton}>Save and finish later</button>
+					<button
+						id="submit"
+						formaction="?/finish"
+						on:click={() => {
+							clearTimeout(debounceTimer);
+							saveButton.disabled = false;
+							saveButton.textContent = 'Save and finish later';
+						}}>Submit application</button
+					>
+				</div>
+			</div>
 		</form>
 
 		<!-- Feedback -->
@@ -269,22 +285,39 @@
 		gap: 0;
 	}
 
-	button {
-		margin-bottom: 0.5rem;
-	}
-
 	#rsvp > * {
 		flex-grow: 1;
-	}
-
-	#submit {
-		margin: 0;
 	}
 
 	#status {
 		border: 2px solid black;
 		padding: 0 1rem;
 		text-align: center;
+		margin-bottom: 1rem;
+	}
+
+	#actions-container {
+		position: sticky;
+		bottom: 0;
+		padding-top: 1rem;
+		background: linear-gradient(transparent, white);
+	}
+
+	#actions {
+		display: flex;
+		flex-direction: row;
+		justify-content: space-between;
+		gap: 0.5rem;
+		position: sticky;
+		padding-bottom: 1rem;
+		background: white;
+	}
+
+	#actions > * {
+		flex: 1;
+	}
+
+	#status button {
 		margin-bottom: 1rem;
 	}
 </style>