+++ src/routes/apply/+page.server.ts
@@ -16,18 +16,8 @@ function formToApplication(questions: Question[], formData: FormData) {
 	// eslint-disable-next-line @typescript-eslint/no-explicit-any
 	const application: Record<string, any> = {};
 	for (const question of questions) {
-		if (
-			question.type === 'SENTENCE' ||
-			question.type === 'PARAGRAPH' ||
-			question.type === 'DROPDOWN'
-		) {
+		if (question.type === 'SENTENCE' || question.type === 'PARAGRAPH') {
 			application[question.id] = formData.get(question.id);
-			// Not necessary, but makes the database a bit cleaner in
-			// that a dropdown either has a valid value or is undefined,
-			// rather than an empty string.
-			if (question.type === 'DROPDOWN' && application[question.id] === '') {
-				application[question.id] = undefined;
-			}
 		} else if (question.type === 'NUMBER') {
 			application[question.id] = Number(formData.get(question.id));
 			if (isNaN(application[question.id])) {
@@ -35,15 +25,23 @@ function formToApplication(questions: Question[], formData: FormData) {
 			}
 		} else if (question.type === 'CHECKBOX') {
 			application[question.id] = formData.get(question.id) === 'on';
-		} else if (question.type === 'MULTISELECT') {
-			const formValue = formData.get(question.id);
-			if (formValue) {
-				try {
-					const parsedValues = JSON.parse(formValue as string);
-					application[question.id] = parsedValues.map((item: { value: string }) => item.value);
-				} catch (error) {
-					console.error('Error parsing JSON:', error);
-				}
+		} else if (question.type === 'DROPDOWN') {
+			const selected = formData.get(question.id) as string;
+			try {
+				application[question.id] = JSON.parse(selected).value;
+			} catch (e) {
+				// Needed because JSON.parse on an empty string errors
+				console.error(e);
+			}
+		}
+		if (question.type === 'MULTISELECT') {
+			const selected = formData.get(question.id) as string;
+			try {
+				application[question.id] = JSON.parse(selected).map(
+					(item: { value: string }) => item.value
+				);
+			} catch (e) {
+				console.error(e);
 			}
 		}
 	}