+++ CHANGELOG.md
@@ -8,6 +8,10 @@
 - Ability for admins to export hacker applications to a CSV
 - Question labels support Markdown
 
+[FIXED]
+
+- Non-hackers showing up on admissions widget
+
 ## v0.4.2
 
 [FIXED]