+++ CHANGELOG.md
@@ -7,6 +7,9 @@
 - Registration questions can now be reordered
 - Ability for admins to export hacker applications to a CSV
 - Question labels support Markdown
+- Multiselect questions
+- Typeahead filtering for dropdowns
+- Users can have multiple roles
 
 [FIXED]